pyproject.toml
[tool.poetry]
name = "deploy-functions-oidc-template"
version = "1.0.0"
description = "Cognite Functions template"
authors = ["hakon.treider@cognite.com"]

[tool.poetry.dependencies]
python = "^3.8"
pytest-custom-exit-code = "^0.3"
cognite-sdk = "^2.17"
cognite-logger = "^0.5"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
pre-commit = "^2.11"

[tool.isort]
profile = "black"
force_single_line = true
atomic = true
lines_after_imports = 2
lines_between_types = 1
line_length = 120
known_third_party = ["cognite","pytest"]

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | tests/.*/setup.py
)/
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
